pyproject.toml
[project]
name = "digitalme"
version = "0.1.0"
description = "Medical data multi-agent processing pipeline"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=1.2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "responses>=0.25.0",
    "httpx",
]

[project.scripts]
digitalme = "main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
